refactor: ディレクトリを細分化して基本構文を basic に移動
