+++ doc/Reference/BasicsSyntax.html
@@ -146,7 +146,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="define_variables-source">
-            <pre><span class="ruby-comment"># File lib/reference/basics_syntax.rb, line 30</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/basics_syntax.rb, line 30</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">define_variables</span>
   <span class="ruby-identifier">integer</span> = <span class="ruby-value">42</span>
   <span class="ruby-identifier">float</span> = <span class="ruby-value">3.14</span>
@@ -191,7 +191,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="manipulate_array-source">
-            <pre><span class="ruby-comment"># File lib/reference/basics_syntax.rb, line 50</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/basics_syntax.rb, line 50</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">manipulate_array</span>(<span class="ruby-identifier">array</span>)
   <span class="ruby-identifier">array</span>.<span class="ruby-identifier">map</span> { <span class="ruby-operator">|</span><span class="ruby-identifier">i</span><span class="ruby-operator">|</span> <span class="ruby-identifier">i</span> <span class="ruby-operator">*</span> <span class="ruby-value">2</span> }
 <span class="ruby-keyword">end</span></pre>
@@ -229,7 +229,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="manipulate_hash-source">
-            <pre><span class="ruby-comment"># File lib/reference/basics_syntax.rb, line 63</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/basics_syntax.rb, line 63</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">manipulate_hash</span>(<span class="ruby-identifier">hash</span>)
   <span class="ruby-identifier">hash</span>.<span class="ruby-identifier">transform_keys</span> { <span class="ruby-operator">|</span><span class="ruby-identifier">key</span><span class="ruby-operator">|</span> <span class="ruby-identifier">key</span>.<span class="ruby-identifier">to_s</span>.<span class="ruby-identifier">upcase</span> }.<span class="ruby-identifier">transform_values</span>(<span class="ruby-operator">&amp;</span><span class="ruby-value">:to_s</span>)
 <span class="ruby-keyword">end</span></pre>
