+++ doc/Reference/Circle.html
@@ -147,7 +147,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="new-source">
-            <pre><span class="ruby-comment"># File lib/reference/class_usage.rb, line 62</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/class_usage.rb, line 62</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">initialize</span>(<span class="ruby-identifier">radius</span>)
   <span class="ruby-keyword">super</span>(<span class="ruby-string">&#39;Circle&#39;</span>)
   <span class="ruby-ivar">@radius</span> = <span class="ruby-identifier">radius</span>
@@ -190,7 +190,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="area-source">
-            <pre><span class="ruby-comment"># File lib/reference/class_usage.rb, line 70</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/class_usage.rb, line 70</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">area</span>
   <span class="ruby-constant">Math</span><span class="ruby-operator">::</span><span class="ruby-constant">PI</span> <span class="ruby-operator">*</span> <span class="ruby-ivar">@radius</span><span class="ruby-operator">**</span><span class="ruby-value">2</span>
 <span class="ruby-keyword">end</span></pre>
@@ -221,7 +221,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="description-source">
-            <pre><span class="ruby-comment"># File lib/reference/class_usage.rb, line 77</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/class_usage.rb, line 77</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">description</span>
   <span class="ruby-node">&quot;Circle with radius #{@radius}&quot;</span>
 <span class="ruby-keyword">end</span></pre>
