+++ doc/Reference/ControlFlow.html
@@ -134,7 +134,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="check_number-source">
-            <pre><span class="ruby-comment"># File lib/reference/control_flow.rb, line 26</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/control_flow.rb, line 26</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">check_number</span>(<span class="ruby-identifier">number</span>)
   <span class="ruby-keyword">if</span> <span class="ruby-identifier">number</span>.<span class="ruby-identifier">positive?</span>
     <span class="ruby-string">&#39;positive&#39;</span>
@@ -177,7 +177,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="day_of_week-source">
-            <pre><span class="ruby-comment"># File lib/reference/control_flow.rb, line 44</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/control_flow.rb, line 44</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">day_of_week</span>(<span class="ruby-identifier">day</span>)
   <span class="ruby-keyword">case</span> <span class="ruby-identifier">day</span>
   <span class="ruby-keyword">when</span> <span class="ruby-string">&#39;Monday&#39;</span>
@@ -221,7 +221,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="is_even-source">
-            <pre><span class="ruby-comment"># File lib/reference/control_flow.rb, line 64</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/control_flow.rb, line 64</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">is_even</span>(<span class="ruby-identifier">number</span>) <span class="ruby-comment"># rubocop:disable Naming/PredicateName</span>
   <span class="ruby-identifier">number</span>.<span class="ruby-identifier">even?</span> <span class="ruby-operator">?</span> <span class="ruby-string">&#39;Even&#39;</span> <span class="ruby-operator">:</span> <span class="ruby-string">&#39;Odd&#39;</span>
 <span class="ruby-keyword">end</span></pre>
