+++ doc/Reference/Functional.html
@@ -134,7 +134,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="apply_function-source">
-            <pre><span class="ruby-comment"># File lib/reference/functional.rb, line 29</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/functional.rb, line 29</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">apply_function</span>(<span class="ruby-identifier">array</span>, <span class="ruby-identifier">func</span>)
   <span class="ruby-identifier">array</span>.<span class="ruby-identifier">map</span> { <span class="ruby-operator">|</span><span class="ruby-identifier">element</span><span class="ruby-operator">|</span> <span class="ruby-identifier">func</span>.<span class="ruby-identifier">call</span>(<span class="ruby-identifier">element</span>) }
 <span class="ruby-keyword">end</span></pre>
@@ -172,7 +172,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="double_lambda-source">
-            <pre><span class="ruby-comment"># File lib/reference/functional.rb, line 40</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/functional.rb, line 40</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">double_lambda</span>
   <span class="ruby-operator">-&gt;</span>(<span class="ruby-identifier">x</span>) { <span class="ruby-identifier">x</span> <span class="ruby-operator">*</span> <span class="ruby-value">2</span> }
 <span class="ruby-keyword">end</span></pre>
@@ -209,7 +209,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="immutable_array-source">
-            <pre><span class="ruby-comment"># File lib/reference/functional.rb, line 52</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/functional.rb, line 52</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">immutable_array</span>(<span class="ruby-identifier">array</span>)
   <span class="ruby-identifier">array</span>.<span class="ruby-identifier">freeze</span>.<span class="ruby-identifier">map</span> { <span class="ruby-operator">|</span><span class="ruby-identifier">x</span><span class="ruby-operator">|</span> <span class="ruby-identifier">x</span> <span class="ruby-operator">+</span> <span class="ruby-value">1</span> }
 <span class="ruby-keyword">end</span></pre>
@@ -246,7 +246,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="pure_addition-source">
-            <pre><span class="ruby-comment"># File lib/reference/functional.rb, line 77</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/functional.rb, line 77</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">pure_addition</span>(<span class="ruby-identifier">a</span>, <span class="ruby-identifier">b</span>)
   <span class="ruby-identifier">a</span> <span class="ruby-operator">+</span> <span class="ruby-identifier">b</span>
 <span class="ruby-keyword">end</span></pre>
@@ -283,7 +283,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="sum_array-source">
-            <pre><span class="ruby-comment"># File lib/reference/functional.rb, line 64</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/functional.rb, line 64</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">sum_array</span>(<span class="ruby-identifier">array</span>)
   <span class="ruby-identifier">array</span>.<span class="ruby-identifier">reduce</span>(<span class="ruby-value">0</span>) { <span class="ruby-operator">|</span><span class="ruby-identifier">sum</span>, <span class="ruby-identifier">x</span><span class="ruby-operator">|</span> <span class="ruby-identifier">sum</span> <span class="ruby-operator">+</span> <span class="ruby-identifier">x</span> }
 <span class="ruby-keyword">end</span></pre>
