+++ doc/Reference/Loops.html
@@ -135,7 +135,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="each_loop-source">
-            <pre><span class="ruby-comment"># File lib/reference/loops.rb, line 63</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/loops.rb, line 63</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">each_loop</span>(<span class="ruby-identifier">array</span>)
   <span class="ruby-identifier">result</span> = []
   <span class="ruby-identifier">array</span>.<span class="ruby-identifier">each</span> <span class="ruby-keyword">do</span> <span class="ruby-operator">|</span><span class="ruby-identifier">item</span><span class="ruby-operator">|</span>
@@ -176,7 +176,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="times_loop-source">
-            <pre><span class="ruby-comment"># File lib/reference/loops.rb, line 79</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/loops.rb, line 79</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">times_loop</span>(<span class="ruby-identifier">times</span>)
   <span class="ruby-identifier">result</span> = []
   <span class="ruby-identifier">times</span>.<span class="ruby-identifier">times</span> <span class="ruby-keyword">do</span> <span class="ruby-operator">|</span><span class="ruby-identifier">i</span><span class="ruby-operator">|</span>
@@ -217,7 +217,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="until_loop-source">
-            <pre><span class="ruby-comment"># File lib/reference/loops.rb, line 45</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/loops.rb, line 45</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">until_loop</span>(<span class="ruby-identifier">limit</span>)
   <span class="ruby-identifier">count</span> = <span class="ruby-value">0</span>
   <span class="ruby-identifier">result</span> = []
@@ -260,7 +260,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="while_loop-source">
-            <pre><span class="ruby-comment"># File lib/reference/loops.rb, line 28</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/loops.rb, line 28</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">while_loop</span>(<span class="ruby-identifier">count</span>)
   <span class="ruby-identifier">result</span> = []
   <span class="ruby-keyword">while</span> <span class="ruby-identifier">count</span> <span class="ruby-operator">&gt;</span> <span class="ruby-value">0</span>
