+++ doc/Reference/OverloadDemo.html
@@ -139,7 +139,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="calculate-source">
-            <pre><span class="ruby-comment"># File lib/reference/class_usage.rb, line 127</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/class_usage.rb, line 127</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">calculate</span>(<span class="ruby-operator">*</span><span class="ruby-identifier">args</span>)
   <span class="ruby-keyword">case</span> <span class="ruby-identifier">args</span>.<span class="ruby-identifier">size</span>
   <span class="ruby-keyword">when</span> <span class="ruby-value">1</span>
