+++ doc/Reference/Rectangle.html
@@ -146,7 +146,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="new-source">
-            <pre><span class="ruby-comment"># File lib/reference/class_usage.rb, line 93</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/class_usage.rb, line 93</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">initialize</span>(<span class="ruby-identifier">width</span>, <span class="ruby-identifier">height</span>)
   <span class="ruby-keyword">super</span>(<span class="ruby-string">&#39;Rectangle&#39;</span>)
   <span class="ruby-ivar">@width</span> = <span class="ruby-identifier">width</span>
@@ -190,7 +190,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="area-source">
-            <pre><span class="ruby-comment"># File lib/reference/class_usage.rb, line 102</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/class_usage.rb, line 102</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">area</span>
   <span class="ruby-ivar">@width</span> <span class="ruby-operator">*</span> <span class="ruby-ivar">@height</span>
 <span class="ruby-keyword">end</span></pre>
