+++ doc/Reference/Shape.html
@@ -154,7 +154,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="new-source">
-            <pre><span class="ruby-comment"># File lib/reference/class_usage.rb, line 27</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/class_usage.rb, line 27</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">initialize</span>(<span class="ruby-identifier">name</span>)
   <span class="ruby-identifier">raise</span> <span class="ruby-constant">NotImplementedError</span>, <span class="ruby-string">&#39;Shape クラスは直接インスタンス化できません&#39;</span> <span class="ruby-keyword">if</span> <span class="ruby-identifier">instance_of?</span>(<span class="ruby-constant">Shape</span>)
 
@@ -194,7 +194,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="area-source">
-            <pre><span class="ruby-comment"># File lib/reference/class_usage.rb, line 38</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/class_usage.rb, line 38</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">area</span>
   <span class="ruby-identifier">raise</span> <span class="ruby-constant">NotImplementedError</span>, <span class="ruby-string">&#39;サブクラスで area を定義してください&#39;</span>
 <span class="ruby-keyword">end</span></pre>
@@ -227,7 +227,7 @@
 
         <div class="method-description">
           <div class="method-source-code" id="description-source">
-            <pre><span class="ruby-comment"># File lib/reference/class_usage.rb, line 47</span>
+            <pre><span class="ruby-comment"># File lib/reference/basic/class_usage.rb, line 47</span>
 <span class="ruby-keyword">def</span> <span class="ruby-identifier ruby-title">description</span>
   <span class="ruby-node">&quot;Shape: #{@name}&quot;</span>
 <span class="ruby-keyword">end</span></pre>
