+++ doc/created.rid
@@ -1,12 +1,12 @@
-Fri, 03 Jan 2025 19:41:45 +0000
-/workspaces/devcontainer-ruby/Gemfile	Fri, 03 Jan 2025 19:32:14 +0000
+Sat, 04 Jan 2025 15:03:50 +0000
+/workspaces/devcontainer-ruby/Gemfile	Sat, 04 Jan 2025 14:40:05 +0000
 /workspaces/devcontainer-ruby/Gemfile.lock	Fri, 03 Jan 2025 19:32:14 +0000
 /workspaces/devcontainer-ruby/LICENSE	Sat, 21 Dec 2024 03:09:47 +0000
-/workspaces/devcontainer-ruby/README.md	Sat, 21 Dec 2024 06:38:12 +0000
-/workspaces/devcontainer-ruby/Rakefile	Sat, 21 Dec 2024 06:29:54 +0000
-/workspaces/devcontainer-ruby/lib/my_project.rb	Fri, 03 Jan 2025 09:30:25 +0000
-/workspaces/devcontainer-ruby/lib/reference/basics_syntax.rb	Fri, 03 Jan 2025 19:36:30 +0000
-/workspaces/devcontainer-ruby/lib/reference/class_usage.rb	Fri, 03 Jan 2025 19:38:36 +0000
-/workspaces/devcontainer-ruby/lib/reference/control_flow.rb	Fri, 03 Jan 2025 19:39:57 +0000
-/workspaces/devcontainer-ruby/lib/reference/functional.rb	Fri, 03 Jan 2025 19:30:46 +0000
-/workspaces/devcontainer-ruby/lib/reference/loops.rb	Fri, 03 Jan 2025 19:41:29 +0000
+/workspaces/devcontainer-ruby/README.md	Sat, 04 Jan 2025 15:03:28 +0000
+/workspaces/devcontainer-ruby/Rakefile	Sat, 04 Jan 2025 14:40:05 +0000
+/workspaces/devcontainer-ruby/lib/my_project.rb	Sat, 04 Jan 2025 14:40:05 +0000
+/workspaces/devcontainer-ruby/lib/reference/basic/basics_syntax.rb	Sat, 04 Jan 2025 14:40:05 +0000
+/workspaces/devcontainer-ruby/lib/reference/basic/class_usage.rb	Sat, 04 Jan 2025 14:40:05 +0000
+/workspaces/devcontainer-ruby/lib/reference/basic/control_flow.rb	Sat, 04 Jan 2025 14:40:05 +0000
+/workspaces/devcontainer-ruby/lib/reference/basic/functional.rb	Sat, 04 Jan 2025 14:40:05 +0000
+/workspaces/devcontainer-ruby/lib/reference/basic/loops.rb	Sat, 04 Jan 2025 14:40:05 +0000
