+++ doc/table_of_contents.html
@@ -89,14 +89,21 @@
     <a href="README_md.html">README</a>
 
     <ul>
-      <li><a href="README_md.html#label-Ruby+-E9-96-8B-E7-99-BA-E7-94-A8+Dev+Container">Ruby 開発用 Dev Container</a>
-      <li><a href="README_md.html#label-E7-89-B9-E5-BE-B4">特徴</a>
-      <li><a href="README_md.html#label-E5-A7-8B-E3-82-81-E6-96-B9">始め方</a>
-      <li><a href="README_md.html#label-1.+-E3-83-AA-E3-83-9D-E3-82-B8-E3-83-88-E3-83-AA-E3-81-AE-E3-82-AF-E3-83-AD-E3-83-BC-E3-83-B3">1. <strong>リポジトリのクローン</strong></a>
-      <li><a href="README_md.html#label-2.+VS+Code+-E3-81-A7-E9-96-8B-E3-81-8F">2. <strong>VS Code で開く</strong></a>
-      <li><a href="README_md.html#label-3.+-E3-82-A2-E3-83-97-E3-83-AA-E3-82-B1-E3-83-BC-E3-82-B7-E3-83-A7-E3-83-B3-E3-81-AE-E5-AE-9F-E8-A1-8C">3. <strong>アプリケーションの実行</strong></a>
-      <li><a href="README_md.html#label-E3-82-AB-E3-82-B9-E3-82-BF-E3-83-9E-E3-82-A4-E3-82-BA">カスタマイズ</a>
-      <li><a href="README_md.html#label-E3-83-A9-E3-82-A4-E3-82-BB-E3-83-B3-E3-82-B9">ライセンス</a>
+      <li><a href="README_md.html#label-Devcontainer+Ruby+Reference">Devcontainer Ruby Reference</a>
+      <li><a href="README_md.html#label-E7-9B-AE-E6-AC-A1"><strong>目次</strong></a>
+      <li><a href="README_md.html#label-E3-82-BB-E3-83-83-E3-83-88-E3-82-A2-E3-83-83-E3-83-97"><strong>セットアップ</strong></a>
+      <li><a href="README_md.html#label-E5-BF-85-E8-A6-81-E3-81-AA-E3-83-84-E3-83-BC-E3-83-AB"><strong>必要なツール</strong></a>
+      <li><a href="README_md.html#label-E7-92-B0-E5-A2-83-E6-A7-8B-E7-AF-89"><strong>環境構築</strong></a>
+      <li><a href="README_md.html#label-Rspec+-E3-81-AE-E4-BD-BF-E3-81-84-E6-96-B9"><strong>Rspec の使い方</strong></a>
+      <li><a href="README_md.html#label-E3-83-86-E3-82-B9-E3-83-88-E3-81-AE-E5-AE-9F-E8-A1-8C"><strong>テストの実行</strong></a>
+      <li><a href="README_md.html#label-RDoc+-E3-81-AE-E7-94-9F-E6-88-90"><strong>RDoc の生成</strong></a>
+      <li><a href="README_md.html#label-E6-89-8B-E9-A0-86"><strong>手順</strong></a>
+      <li><a href="README_md.html#label-Reference+-E4-B8-80-E8-A6-A7"><strong>Reference 一覧</strong></a>
+      <li><a href="README_md.html#label-RuboCop+-E3-81-AE-E5-88-A9-E7-94-A8-E6-96-B9-E6-B3-95"><strong>RuboCop の利用方法</strong></a>
+      <li><a href="README_md.html#label-E3-82-B3-E3-83-BC-E3-83-89-E3-82-B9-E3-82-BF-E3-82-A4-E3-83-AB-E3-83-81-E3-82-A7-E3-83-83-E3-82-AF"><strong>コードスタイルチェック</strong></a>
+      <li><a href="README_md.html#label-E8-87-AA-E5-8B-95-E4-BF-AE-E6-AD-A3"><strong>自動修正</strong></a>
+      <li><a href="README_md.html#label-E7-89-B9-E5-AE-9A-E3-83-95-E3-82-A1-E3-82-A4-E3-83-AB-E3-82-92-E3-83-81-E3-82-A7-E3-83-83-E3-82-AF"><strong>特定ファイルをチェック</strong></a>
+      <li><a href="README_md.html#label-E3-83-A9-E3-82-A4-E3-82-BB-E3-83-B3-E3-82-B9"><strong>ライセンス</strong></a>
     </ul>
   </li>
   <li class="file">
