+++ lib/reference/basic/class_usage.rb
