+++ lib/reference/basic/functional.rb
